removed libraries, rewrited cmakelists.txt for automatic find sfml, fixed game field coord, fixed eventlistener and added WindowMutex for sync gui and event threads sf::Window usage

+++ tank.hpp
@@ -18,8 +18,8 @@ namespace baho
     void setPosition(point_t begin);
     void setSprite(const sf::Sprite & sprite);
     Bullet shot(Bullet prototype);
-    double getX() const;
-    double getY() const;
+    float getX() const;
+    float getY() const;
     void destroy();
     bool isDestroyed() const;
     bool mayDelete() const;
